requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
supabase>=2.4.0
Pillow==10.2.0
aiohttp==3.9.3
asyncpg==0.29.0
APScheduler==3.10.4
python-telegram-bot==20.8
